separated the implementation from the sample form
