+++ components/reusable-components/formik/FormikSample.module.scss
@@ -1,34 +1,16 @@
-.formContainer {
-  border: 2px solid white;
-  max-width: 600px;
+.field {
+  position: relative;
 
-  .form {
-    display: flex;
-    flex-direction: column;
-    padding: 1rem;
-    gap: 1.7rem;
-
-    button {
-      background-color: white;
-      height: 2rem;
-      font-weight: 600;
-    }
+  input {
+    all: unset;
+    border: 1px solid white;
+    padding: 0 1rem;
+    height: 2.5rem;
   }
-
-  .field {
-    position: relative;
-
-    input {
-      all: unset;
-      border: 1px solid white;
-      padding: 0 1rem;
-      height: 2.5rem;
-    }
-    .error {
-      position: absolute;
-      left: 0;
-      top: 100%;
-      color: red;
-    }
+  .error {
+    position: absolute;
+    left: 0;
+    top: 100%;
+    color: red;
   }
 }
